+++ js/src/admin/index.js
@@ -1,6 +1,6 @@
 import { extend } from 'flarum/extend';
 import addPermissions from "./addPermissions";
 
-app.initializers.add('flagrow-subscribed',() => {
+app.initializers.add('fof-subscribed',() => {
     addPermissions();
 });