+++ js/src/admin/permissions/discussionCreated.js
@@ -1,7 +1,7 @@
 export default function (items) {
     items.add('subscribeDiscussionCreated', {
         icon: 'fas fa-bell',
-        label: app.translator.trans('flagrow-subscribed.admin.permission.subscribe_to_discussion_created'),
+        label: app.translator.trans('fof-subscribed.admin.permission.subscribe_to_discussion_created'),
         permission: 'subscribeDiscussionCreated'
     }, 95);
 }