+++ js/src/forum/notifications/DiscussionCreatedNotification.js
@@ -13,6 +13,6 @@ export default class DiscussionCreatedNotification extends Notification {
     }
 
     content() {
-        return app.translator.trans('flagrow-subscribed.forum.notifications.discussion_created_text', {user: this.props.notification.fromUser()});
+        return app.translator.trans('fof-subscribed.forum.notifications.discussion_created_text', {user: this.props.notification.fromUser()});
     }
 }